fix: do not add services if Spanner dialect is not used
The Spanner dialect would add services and configuration, even if the Spanner
dialect was not used, but was present on the classpath. This makes it impossible
to develop an application that can run on both Cloud Spanner and other databases,
as the Spanner services would be loaded and used for all databases.

It also made it impossible to use for example H2 for testing purposes, while
using Cloud Spanner for production.