fix: close session and sessionFactory when finished
The sample application should close the session and session factory when
the application finishes. Failing to do so, will keep at least one JDBC
connection open, which again could cause sessions to be kept around on the
backend until they are garbage collected one hour later.

This chanage also turns off the cleanup of daemon threads by the maven-exec
plugin. Otherwise it reports a large number of false positives of threads
that are supposedly not shutdown. They are however closed by a shutdown hook
after the check has been executed.