fix: do not ignore exception during schema migration
The Cloud Spanner Hibernate dialect will include the statements
`START BATCH DDL` and `RUN BATCH` as auxiliary database objects. This
ensures that the entire schema migration is executed as a single DDL
batch, which both improves the execution speed of the statements, as
well as makes the parsing of all statements atomic. However, Hibernate
will ignore any SQLException that is thrown when creating an auxiliary
database object. This means that if the `RUN BATCH` statement returns a
SQLException, for example because one of the DDL statements is invalid,
Hibernate will just silently ignore the error and proceed as if the
schema migration worked.

Fixes #470