feat: support bit-reversed sequences
Adds support for bit-reversed sequences. These are recommended above
normal, monotonically increasing sequences for Cloud Spanner, as using a
monotonically increasing value for a primary key can cause write hot
spots.

This change also fixes the initialization of the backing tables for
these sequences. Normally, the Cloud Spanner Hibernate dialect would not
insert a value into the table that is backing a sequence, unless it was
the default `hibernate_sequence` table. This would mean that generators
that used custom table would be generated, but initialized, which would
mean that they would not work out of the box.