feat: add interceptor that enables Mutation usage
Adds an interceptor that can be used in combination with the Spanner
Hibernate dialect to use Mutations instead of DML. Mutations can have a
performance advantage compared to DML, but with the drawback that
mutations cannot be read during the same transaction. This interceptor
therefore by default only uses mutations when a flush is initiated by
Hibernate because the transaction is committing. In all other cases the
interceptor will remain inactive and Hibernate will continue to use DML.