feat: support bit-reversed sequences on the emulator
Bit-reversed sequences require the use of a read/write transaction.
Hibernate starts a second read/write transaction when it needs to
fetch new identifiers, and this transaction is aborted when it is
running on the emulator. The PooledBitReversedSequenceGenerator
now solves this by rolling back that transaction after getting the
sequence values. This is semantically equivalent to committing the
transaction, as the state of the sequence is durably changed
regardless whether the transaction is committed or rolled back.

Fixes #983