feat: transaction tagging (#977)
* feat: transaction tagging

Adds support for transaction tags using an interceptor that
traverses the call stack up to the method that actually
started the transaction. The tag is then determined by:
1. Looking for an @TransactionTag annotation on the method.
2. Or, if enabled, automatically generating a tag from the
   class name and method name.

The option to automatically add tags based on class and method
name can be enabled through a system property. This makes it
possible to enable this feature only during debugging, and
disable it again when it is no longer needed.

* chore: cleanup