fix: use unnest instead of array_contains (#1552)
Hibernate by default uses the array_contains function when a predicate
is created to search for entities where the ID is in an array of elements.
array_contains is however not a supported function in Spanner GoogleSQL,
and also using `in unnest(@values)` is more efficient on Spanner.

The `visitInArrayPredicate` method is therefore now overridden in the
Spanner dialect and uses an `in unnest(@values)` construct instead.

Fixes #1546